+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `only_extrema` parameter to `EpanetBinHandler.export_to_database` to skip time series inserts and only export aggregated stats.
+
 ## [0.4.1] - 2026-06-23
 
 ### Fixed